+++ CHANGELOG.md
@@ -67,7 +67,7 @@ In order to build the DUNE core modules you need at least the following software
   https://gitlab.dune-project.org/core/dune-common/-/merge_requests/1251
   for more details on how to use this feature in the source modules and on
   some new feature.
-  
+
 - Remove deprecated `add_directory_test_target` function.
 
 - Do not overwrite the `add_test` cmake function with an error message.
@@ -139,6 +139,9 @@ In order to build the DUNE core modules you need at least the following software
   `std::array` interfaces.
 
 ## Build System
+
+- Introduce a new cmake function `dune_export_toolchain` to export system and compiler
+  configurations into a toolchain file.
 
 - Improve the the function `dune_add_library` by separating the target types normal, interface, and
   object. Additional properties can be passed to the cmake function like `LINK_LIBRARIES`,